Using numba in get_slices
